some new entities

+++ vex-api/src/main/java/com/vex/repositories/ProductRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends ReactiveCrudRepository<Product, Long> {
+public interface ProductRepository extends ReactiveCrudRepository<Product, Integer> {
 
 }
